getframe: Use a fixed stride for dsbuf.
In a multi-segment record, the input frame size (framelen) may change
from one segment to another.  If framelen is used as the stride for
the deskewing buffer, and the size of the buffer is not a multiple of
framelen, this can result in overflowing the buffer as well as
garbling the results.

In getframe, use tspf (which is constant throughout the record, and
always greater than or equal to framelen) in place of framelen.

When calculating dsblen initially, use tspf instead of framelen (the
two should be equal at that point, but this reduces ambiguity.)

Fixing the deskewing behavior so that it actually produces *correct*
results, in general, for variable-layout records, will be done in a
later version.