WPG: document return value of -3 from isigopen.
For most applications, if isigopen fails, the precise reason isn't
important.  However, for applications that open multiple records
simultaneously (using isigopen("+rec", ...)), it may be important for
the application to know whether previously-opened signals have been
closed or left open.

(Ideally we wouldn't ever close existing signals, but in some cases it
may be impractical or even impossible to "roll back" to the previous
state.)