isigopen: correctly limit number of signals in '+' mode.
The 'nsig' parameter to isigopen is the maximum number of new signals
that are to be opened, and the size of 'siarray', as provided by the
caller.

(As the WPG puts it, the return value of isigopen ought to be "never
greater than NSIG"; this is also the way existing applications such as
'nst' expect it to behave.)

In previous versions, however, 'nsig' was set to 'nsig+nisig', but was
still interpreted as a limit on the number of newly-opened signals.
Thus, in "plus" mode, isigopen could open a larger number of signals
than the application requested, overflowing 'siarray' and possibly
resulting in other buffer overflows.