isigopen: allocate correct number of signal groups.
If we end up skipping some signal groups (either because the file
can't be read or because the number of signals exceeds nsig), we may
end up opening a larger total number of files than expected.

It's tricky to figure out the exact maximum number of groups we might
open, but we know it cannot be greater than the maximum number of
signals (nsig), and it cannot be greater than the number of groups
specified by the header (hsd[navail-1]->info.group + 1).

(In practice, it seems unlikely that the overhead of allocating igdata
will ever make any difference.)