wfdb_fopen: don't overrun buffer.
wfdb_fopen checks to see whether the filename contains the substring
"://", and treats it as a URL if so.

However, the code did not properly check to see whether it had reached
the end of the string.  For example, if a signal file name ends with a
colon, wfdb_fopen might work correctly, might fail, or might crash,
depending on the contents of memory following the end of the input
string.

A side effect of fixing this bug is that a signal file name such as
"a:://b" will now be parsed as a URL (and rejected, since it is not a
valid URL.)