setmsheader: handle segments longer than LONG_MAX frames.
If a single segment is longer than LONG_MAX frames, then readheader
will set hsd[0]->info.nsamp to zero (indicating "unknown length").  To
obtain the true length of the segment in this case, we must use
nsamples instead, which is a WFDB_Time.  However, readheader will only
set nsamples if it is initially zero.