Use -static-libgcc when building the shared library on mingw.
If this option is not used, current versions of mingw-w64 will link to
a shared libgcc for the __divdi3 and __moddi3 (long long division and
modulus) functions.  We want to avoid the use of an extra DLL in order
to keep installation simple.

(Note that this only seems to affect DLLs; when linking an executable,
these functions are statically linked by default.  Thus, this option
should not be used when linking the WFDB applications with either a
shared or static libwfdb.

Indeed, when linking C++ applications or libraries, it is somewhat
risky to use static-libgcc if you don't know what you're doing.)