getann, ungetann wrappers: copy or clear padding appropriately.
If the application does things like storing annotations in a hash
table, it may be important that the structure is initialized and
copied in the same way as past library versions have done.

If getann is called following ungetann, the structure returned is
copied from the original structure passed to ungetann.  If getann is
called at other times, any padding bytes in the structure are set to
zero.