Add option to run tests without network access.
For both libcheck and appcheck, if the environment variable
WFDB_NO_NET_CHECK is set to a non-empty value, perform NETFILES tests
using a file:// URL instead of the default http:// URL.

(For libcheck, only udb/100s is required, which was already included
in the package.  For appcheck, a small segment of record
mimicdb/237/237 is now included in the checkpkg/input directory.)

This allows testing the package in a fully reproducible manner,
without relying on the PhysioNet web server or any other web server
being accessible at the time the package is built.

For most purposes, this should be adequate to test that the WFDB
library itself is functioning properly, since libcurl takes care of
ensuring that different protocols behave similarly from the caller's
point of view.

However, there are inevitably some differences between the protocols
(for example, file:// URLs don't have headers such as Content-Range),
so when making changes to the NETFILES implementation, it is still
wise to test extensively against real HTTP(S) servers.

For users installing WFDB from source, we still perform HTTP tests by
default, so that users can see whether libcurl is correctly installed
and configured to allow access to PhysioBank.