isgsetframe: avoid internal overflows.
Avoid overflows in the calculation of 'b' by checking the number of
samples per frame.  This is a fixed limit for each format (UINT_MAX/2
samples per frame for format 16, UINT_MAX/3 samples per frame for
format 212, etc.)

To avoid confusion, also replace the use of an 'int' counter with an
'unsigned int' in formats 212, 310, and 311.