get_ann_table: handle types without descriptions.
In an annotation modification label, if the mnemonic was not followed
by whitespace, the previous code was trying to read past the end of
the aux string.  (Sometimes this might actually work as desired, since
auxbuf is initially filled with zeroes; but in many cases this would
result in garbage.)

Note that prior to version 10.6.0, when writing an annotation file,
the library would always include a space after the annotation mnemonic
and before the description.  In version 10.6.0 and 10.6.1, this
behavior changed: if the annotation description was undefined, the
library would omit both the description and the trailing space.

This code also avoids the use of strtok() and avoids two memory leaks.