put_ann_table: treat null description as empty.
If anndesc() returns null, include a space after the mnemonic just as
if it returned an empty string.

Note that older versions of WFDB (before 10.6.0) always included the
space (although they then proceeded to pass a null pointer to sprintf,
invoking undefined behavior), and omitting the space tickles a bug in
previous versions of get_ann_table().

Moreover, note that it is impossible to set an annotation description
to null; calling setanndesc(i, NULL) actually sets the description to
the empty string.  So treating null as being equivalent to the empty
string is more consistent and parsimonious anyway.