sigmap_init: allow fixed-layout with undefined nsamp.
Multi-segment records can be either "fixed-layout" or
"variable-layout".  Variable-layout mode is defined by the first
segment having a length of zero.

However, isigopen may set nsamp to zero even if the segment is
non-empty, for various reasons:
- the segment header doesn't specify the length
- the signal checksums are missing
- the length is greater than LONG_MAX

Databases meant to be portable should avoid these cases, but they can
happen, so we need to handle them reasonably.

It seems more robust, in general, for fixed/variable-layout mode to be
controlled by the master header.  So instead of looking at the nsamp
from the first segment header, look at the length of the first segment
as defined in the master header.