SUALLOC, SREALLOC: consistent behavior for size 0.
The behavior of calloc() or realloc() when the requested size is zero
is implementation-defined.  On many common platforms:

- calloc(0, 0) returns a unique, non-null pointer;

- realloc(0, 0) returns a unique, non-null pointer;

- realloc(p, 0), for p != 0, frees p and returns a null pointer.

Thus, SALLOC(p, 0, 1) would typically succeed (and set p to a non-null
pointer).  On the other hand, SREALLOC(p, 0, 1) would either succeed
or "fail" depending on the initial value of p.

Thus, existing applications might rely on being able to use SALLOC,
SUALLOC, or SREALLOC with an initially-null pointer, with zero size
(because this works on common platforms.)  The application might
further rely on the fact that if the macro succeeds, the pointer is
non-null.

We may assume that existing applications don't rely on being able to
use SREALLOC as a substitute for SFREE (because, on common platforms,
this raises an error - and if it *doesn't* raise an error, that means
that the block has not actually been freed.)

Change the behavior of these macros to be consistent:

- Either SUALLOC(P, N, S) or SREALLOC(P, N, S) always attempts to
  allocate at least one byte of memory.

- If it succeeds, P is set to a non-null pointer.

- If it fails, P is set to null and MEMERR is invoked.