wfdbsetiskew: check bounds of skew argument.
If the skew argument is out of bounds, do nothing.  Previously,
wfdbsetiskew would cause the library to crash or return garbled
results if the skew argument was out of bounds.

Skew values must be between 0 and (dsblen/tspf)-1 (which should
normally equal skewmax, but this is not always the case in
multi-segment records.)