wfdb_parse_path: allow colons in URL authority.
Colons may be used within the "authority" portion of a URL:
- to separate username from password (http://root:letmein@localhost/)
- to separate host from port (http://localhost:8080/)
- to separate components of an IPv6 literal address (http://[::1]/)

In particular, using a non-default port number is often useful for
testing and for private database servers.

On the other hand, historically, WFDB has allowed colons to be used to
separate components of the database path (for example,
WFDB="http://physionet.org/physiobank/database:/usr/database").
Changing this behavior would be risky as it might result in private
data being sent over the network.

Thus, interpret colons as part of a URL only if they occur within the
authority portion (between the second and third slashes.)