Add .exe when needed on mingw and cygwin.
The Cygwin and MinGW compilers and binutils will automatically add
".exe" to the names of executable files.  For example, on Cygwin:

   "gcc hello.c -o hello" creates "hello.exe"
   "strip hello" modifies "hello.exe"

In contrast, *cross*-compilers historically have not had this
behavior, so using (for example) the cross-compilers packaged in
Debian stretch and earlier:

   "x86_64-w64-mingw32-gcc hello.c -o hello" creates "hello"
   "x86_64-w64-mingw32-strip hello" modifies "hello"

Recently, this behavior has been changed in GCC so that it now matches
the native Cygwin/MinGW behavior; e.g. on Debian buster:

   "x86_64-w64-mingw32-gcc hello.c -o hello" creates "hello.exe"

However, binutils still behaves in the traditional way:

   "x86_64-w64-mingw32-strip hello" modifies "hello"

In order to work correctly using both old and new toolchains (and make
testing more consistent between native and cross environments), it's
best to explicitly include the executable suffixes.

To do this, we define a variable 'EXEEXT' in the .def file, which is
.exe on Cygwin/MinGW and blank everywhere else.