pschart, psfd: avoid integer overflows in averaging samples.
pschart calculates the mean value of each strip in order to
automatically position it on the page.  psfd calculates the mean of
the decimated samples, which are each calculated by averaging a
(potentially large) number of input samples.

These calculations are performed on WFDB_Sample values (which may be
as large as 32 bits, for format 32), and the sum was stored as a long,
so it could easily overflow if there were a large number of samples
(nsamp / nisamp / decf) or if the signals were high resolution.

Instead, avoid this issue by calculating the sums as doubles.