Correct types in Fortran examples and documentation.
Some WFDB functions, and the corresponding Fortran wrapper functions,
return floating-point results:

- getifreq
- getafreq
- getiafreq
- getiaorigfreq
- aduphys
- sampfreq

while other functions have floating-point parameters:

- getsiginfo
- setsiginfo
- setifreq
- setafreq
- setiafreq
- physadu
- getcal
- putcal
- setsampfreq
- setcfreq
- setbasecount

Despite documentation and examples to the contrary, the de-facto API
for these has always been to pass and return floating-point values as
"double precision" ("double" in C), rather than "real" ("float" in C).
Indeed, ecgpuwave relies on this in the cases of sampfreq and
getsiginfo.

(On x86-32, and on many older systems, a function that is declared as
returning a "float" actually uses a double-precision register, which
may have hidden this problem.  However, this is not the case on
x86-64.)