SUALLOC: wrap in do { ... } while (0).
This ensures that the macro operates syntactically as a statement
rather than a block, so it can be used in places where a statement is
required.

One place that actually relied on SUALLOC being a block rather than a
statement was the definition of SALLOC.  Hopefully applications have
not followed this example!