getskewedframe: Avoid underflowing nsamp.
is->info.nsamp is a long (for compatibility), not a WFDB_Time.
Consequently, if the record is longer than LONG_MAX frames, nsamp is
initially set to zero.

If the number of frames exceeds -LONG_MIN, then decrementing nsamp
eventually underflows (which is undefined behavior.)  On typical
systems, it will wrap around and become positive again, leading to a
spurious EOF error (if the length of the record is between -LONG_MIN
and LONG_MAX-LONG_MIN) or a spurious checksum error (if the record is
longer than LONG_MAX-LONG_MIN).

Avoid these problems by avoiding decrementing nsamp once it becomes
negative.