WFDB_FMT_LIST: add formats 508, 516, and 524.
These formats will be used to designate FLAC signal files.  Format 508
indicates 8-bit, format 516 indicates 16-bit, and format 524 indicates
24-bit resolution.  Other format codes between 500 and 532 (inclusive)
are reserved.

(Although the FLAC format permits resolutions from 4 bits to 32 bits,
libFLAC currently does not support more than 24 bits.  Furthermore,
many other programs and libraries are limited to multiples of 8 bits.
The overhead of encoding, say, a 12-bit signal as a 16-bit FLAC stream
is usually quite small.)

FLAC files are self-describing, so the format code is not strictly
necessary to identify the format of an input file.  However, the
format code determines the sample resolution when writing an output
file, as well as the sentinel value that is used to represent
WFDB_INVALID_SAMPLE.