sigmap_init: handle FLAC formats.
When reading a segment of a variable-layout record, we need to know
the minimum and maximum sample value permitted by the format in order
to use the "fast path" for sample scaling.

For FLAC files in format 508, 516, or 524, the sample range is the
same as for formats 80, 16, or 24 respectively.  (It is an error for
the FLAC stream resolution to exceed the resolution specified by the
format code.)