nf_fseek: check for overflows and negative values.
If the result of nf_fseek would be a negative file position or an
overflow, this should be treated as an error, as robust
implementations of fseek will do.

There is no need to check for overflows in the SEEK_SET case, but
negative values should be treated as an error.

In the SEEK_CUR case, we want to catch both negative values less than
-nf->pos, and positive values greater than (nf->cont_len - nf->pos).
Casting nf->pos to unsigned long and then adding the offset should
handle both cases: in the first case the result wraps around (and must
be at least (unsigned long)LONG_MIN, which is greater than LONG_MAX)
and in the second case the result is at most 2*LONG_MAX, so it cannot
wrap.

In the SEEK_END case, we want to catch negative values less than
-nf->cont_len, and all positive values; this avoids the need to check
for overflows.