Implement FLAC signal reading.
When WFDB_FLAC_SUPPORT is defined, implement the functions
flac_getsamp(), flac_isopen(), flac_isclose(), and flac_isseek().

Decoding signals with libFLAC uses a push model; each time we call
FLAC__stream_decoder_process_single(), the library decodes some
arbitrarily defined block of samples, and invokes a callback function
(iflac_samples()).  Since the buffers passed to the callback function
may not be valid after the function returns, we need to copy the
contents into a persistent buffer (ig->buf) until the application is
ready to read them.

While we're at it, it's useful for iflac_samples() to rearrange the
samples into WFDB frame order, so that flac_getsamp() simply needs to
read a value and advance a pointer.  For simplicity, the buffer is
stored as an array of FLAC__int32, regardless of the input resolution.

iflac_samples() also needs to accommodate a few eccentricities of the
FLAC library and format:

- We need to reject streams that contain the wrong number of signals
  or the wrong resolution.

- If the input stream is invalid, libFLAC will sometimes try to fill
  in the gap with zeroes; we want to report the error to the
  application, rather than returning incorrect data.

- To work around issues with FLAC__stream_decoder_seek_absolute(), we
  want to discard the first sample of each signal, after a seek.

In addition to iflac_samples(), we need to define the functions
iflac_read(), iflac_seek(), iflac_tell(), iflac_length(), and
iflac_eof(), to allow a WFDB_FILE to be used as a source of input.
Finally, we need to define a function iflac_error() that will be
called in case of an invalid input stream.