[WIP] getskewedframe: handle FLAC formats.
For FLAC signal files, the actual work of reading a sample will be
handled by flac_getsamp().

If the input stream is corrupted and cannot be decoded, flac_getsamp()
will display an appropriate error message and set ig->stat to -2;
getskewedframe() itself should not display an error message in that
case.