wfdbmap: handle signal files that can't be opened.
If the record length is unspecified but the signal format is known, we
try to determine the length from the size of the signal file.  If the
file cannot be found (wfdbfile fails) or cannot be opened (fopen
fails) or is not a regular file (fseek fails) then this will not work;
the length should be left unset, which will result in an error if the
length cannot be determined by other means.