setinfo: report output errors when closing a file.
When closing an info file (either when wfdbquit is called, or setinfo
is called a second time), if either the file error flag is set before
calling fclose, or fclose itself returns a nonzero value, display an
error message.

To allow the application to detect failures, it can call
setinfo(NULL), which will now return a negative value if it is unable
to write the previously-opened info file.