annopen: report output errors when closing a file.
When closing an output annotation file (either when wfdbquit or
oannclose is called, or annopen is called without a "+" argument), if
either the file error flag is set before calling fclose, or fclose
itself returns a nonzero value, display an error message.

To allow the application to detect failures, it can call
annopen("", NULL, 0), which will now return a negative value if it is
unable to write one or more of the annotation files, or if sortann
fails, or if the output was out of order and sorting is disabled.