osigfopen: report output errors when closing a file.
When closing a signal file (either when wfdbquit is called, or
osigopen or osigfopen is called a second time), if either the file
error flag is set before calling fclose, or fclose itself returns a
nonzero value, display an error message.

Additionally, in the case of stdout (which is not closed), call fflush
before checking the error flag.

To allow the application to detect failures, it can call
osigfopen(NULL, 0), which will now return a negative value if it is
unable to write one or more previously-opened signal files.