Implement WFDB_FLAC_CHECK_MD5 environment variable.
The FLAC stream header contains an MD5 hash of the original data.  In
most cases it's a waste of time for WFDB applications to try to
calculate the hash, but it may sometimes be useful to verify the
file's internal correctness.

If the environment variable WFDB_FLAC_CHECK_MD5 is defined with a
non-empty value, then enable hash checking in libFLAC.  There isn't
any good way to report success or failure to the application, since
libFLAC only checks the hash when the stream is closed, which can only
happen during wfdbquit() or a subsequent isigopen().  So we simply
display an error message in this case.

Note that hash checking will only work if the application reads the
entire record from beginning to end, and closes the record when
finished.  Most applications do not do this.