rgetvec: round to nearest integer in low-resolution mode.
In low-resolution mode, rgetvec returns one sample per signal per
frame, which equals the mean of the input samples.  To avoid biasing
the result towards zero, the result should be rounded to the nearest
integer.

For example, in the first frame of mimicdb/041/, we have:

signal    samples            mean      old    new
III       -59 -60 -60 -60    -59.75    -59    -60
I         265 287 265 265    270.50    270    271
V          23  23  23  23     23.00     23     23