struct iadata: increase buffer size to avoid clobbering aux.
getann() always reads ahead and tries to parse the annotation
following the one that is to be returned to the caller.  This means
that there must be space to store the current annotation's aux string
(which will be returned to the caller) as well as the next
annotation's aux string (which we are reading right now.)  An aux
string may be as large as 257 bytes when counting both the length byte
and the extra '\0' added at the end.

Although a 520-byte buffer would be large enough to hold any two aux
strings (indeed, a 514-byte buffer would suffice), getann() does not
simply alternate between two halves of the buffer; rather, it stores
each aux string starting at the next unused address in the buffer, and
restarts at zero once there is not enough room left to fit the next
string.

(A side effect of this strategy is that if the annotation file only
contains fairly short aux strings, it may be possible for an
application to naively read multiple annotations in a row without
making copies of the aux data.  Although this is unwise, it's possible
that such applications exist.)

As a result, with previous versions of WFDB, if there were two long
aux strings in a row, one could partially overwrite the other, causing
the application to see garbled data.  Whether this would occur would
depend on the precise lengths of the aux strings as well as the aux
strings that had previously been read.  For example, if an annotation
file contained one annotation with a 128-byte aux string, followed by
two annotations with 255-byte aux strings, the second annotation's aux
string would be garbled when trying to read it.

(An application would be able to *write* such an annotation file
without problems - unless the annotations were written out of order,
in which case sortann would corrupt them.)

If this bug were triggered, then following a call to getann:

- The contents of aux would be incorrect, and would instead contain
  some portion of the following annotation's aux string.

- aux[0] usually would not equal strlen(aux+1).

- There would be at least one '\0' somewhere between aux[1] and
  aux[256] - the string, though incorrect, would still be
  zero-terminated.

- If the application interpreted aux[0] as the length of the string
  (for example, if it passed the resulting annotation to putann())
  then it might read past the end of the real aux data to whatever was
  stored in the buffer beforehand - either zeroes or the aux data of
  some even older annotations.  However, in no case could the
  application be tricked into reading a byte later than
  auxstr[255+UCHAR_MAX].

Because the annotation file is read sequentially from the beginning,
and the buffer is initialized to all zeroes, the garbled result
(whether interpreted as a zero-terminated C string or a
length-prefixed Pascal string) would in fact be deterministic.
However, if an application were to read an annotation file, then
rewind to the beginning using iannsettime(), then read the file again,
the results might be different the second time.

To avoid these problems, increase the size of auxstr to 771 bytes.
This means that if the condition at line 535 is satisfied:

- ia->index >= AUXBUFLEN - 2 - len

- therefore, since len <= 255, ia->index >= 771 - 2 - 255 == 514

- ia->auxstr + ia->index == annot->aux + 2 + annot->aux[0]

- therefore, since annot->aux[0] <= 255,
  annot->aux - ia->auxstr >= ia->index - 2 - 255 >= 514 - 257 == 257

- therefore, writing at most 257 bytes at ia->auxstr[0] cannot
  overwrite any part of annot->aux.