getwfdb: do not alter environment variables.
Previously, getwfdb would invoke setwfdb.  setwfdb stores the new WFDB
path, and other settings, in environment variables.

Altering environment variables is generally not thread-safe (it's not
merely unsafe to call putenv concurrently from two threads; it's
unsafe to call putenv while another thread is calling getenv), which
meant it was dangerous to call a function like isigopen if there were
any other threads running in the same process.

It also doesn't make sense for isigopen to alter the behavior of
future child processes (for example, if we later invoke another
program that was linked with a different version of WFDB, with a
different default database path.)

Instead, ensure that the environment is only modified when the
application explicitly calls setwfdb().