setwfdb: read path from "@" files.
If the WFDB environment variable is set to "@foo", this indicates we
should use the contents of the file "foo" as the database path.  This
behavior should be consistent regardless of whether the program calls
getwfdb() and/or setwfdb(NULL) before calling isigopen().

Previously, getwfdb() would handle "@" files, but setwfdb() would not.
(However, setwfdb() would store the given string in the environment,
where it might be used by future child processes.)

For consistency, we want setwfdb("@foo") (or setwfdb(NULL), if $WFDB
already equals "@foo") to set $WFDB to "@foo".  But - also for
consistency - we want getwfdb() to return the actual database path
that the library will use (i.e., the contents of the file "foo".)

So setwfdb should call wfdb_export_config before calling
wfdb_getiwfdb, but then call wfdb_parse_path afterwards (and set
wfdbpath to the expanded path).