Skip testing the interactive output of xform.
The interactive output of xform will vary depending on whether it is
compiled with FLAC support or not.  It's not especially important for
the test suite to verify that the interactive output is exactly as
expected, so remove this part of the test.  The important part of this
test is to check that the output files produced by xform are correct.