Avoid memory leaks in wfdb_getiwfdb.
The internal function wfdb_getiwfdb is called in order to replace an
indirect path ("@foo") with the actual contents of the file "foo" (or
the file that "foo" points to, etc.)

Since this requires allocating a new buffer, this would leak memory.
Instead, change the function to take a 'char **' argument, which must
point to an existing malloc-allocated buffer.  That buffer will be
freed if necessary and replaced with a newly allocated buffer.

There's no reason for this function to be externally visible, so make
it static at the same time.