edfparse: handle non-zero-terminated text fields.
Fields in the EDF header are stored as fixed-length character arrays
padded with spaces.  When parsing each field, we want to ignore these
trailing spaces, and always zero-terminate the result.  The previous
version of the function would do this for some fields, but not
systematically.

For example, if the signal unit string was eight characters long, the
previous version of the function would not add a zero terminator, and
the units string returned by isigopen would contain unpredictable
garbage.

Similarly, many other fields (recording date, recording time, header
size, number of frames, frame duration, physical maximum/minimum,
digital maximum/minimum, samples per frame) could be interpreted
incorrectly if the value was eight characters long and the preceding
field did not contain a space character in the ninth position.