wfdb2mat: probe length of record if unspecified.
In order to generate the mat file prolog (and check if the size
exceeds the limits of the format) the input record length must be
known.  This is true even if an explicit "-t" or "-l" option is given,
since the end time is clamped to the record length.

If the input header file has no length specified, we must guess based
on the size of the signal file.  To do so without relying on the
specific signal file format, use a binary search to find the smallest
value of t for which either isigsettime or getvec fails.