snip: fix multi-segment format with unknown base date.
If the input record has a base time but not a base date, and the
starting time occurs later than the first day of the record, then
mstimstr(-from) will include a day number but not a date.

This day number will be ignored when the time string is passed to
setbasetime, and setheader will simply write the base time with no
base date.

However, in the case of multi-segment output, snip constructs the
master header file by hand.  Therefore, the day number (which is not
valid as a base date in a header file) must be removed in this case.