signum: fix format of output.
signum is intended to print a tab-separated list of matching signal
numbers, or "X" if there are no matching signals.

Instead, it would previously print only the *first* matching signal
number, preceded by N tab characters, and followed by an "X" if N was
zero, which was not very useful.