configure: use 'command -v' instead of 'which' if possible.
'which' has never been standardized, and gives slightly different
results on different systems.  POSIX provides 'command -v' as a
standardized replacement for 'which'.

When configuring the package, check whether 'command -v' appears to be
supported, and select one of the two wrapper scripts 'which-c' or
'which-w' to use when searching for compilers and other programs.

Furthermore, make these checks more robust by use 'test -x' to check
whether the result of 'command -v' or 'which' is actually an
executable program.  In the case of 'which', ignore the standard error
rather than mixing it with standard output.