getskewedframe: seek correctly at the start of a signal file.
If a signal file contains a prolog, then after the signal file is
first opened and before reading the first frame of data, we must skip
to the actual start of data (normally by calling wfdb_fseek, sometimes
by calling wfdb_fread.)

In principle, this could be done by isigopen, but that could be a
waste of time if the application intends to call isigsettime
immediately or if it doesn't intend to read signal data at all.

Previously, therefore, getskewedframe would invoke isgsetframe if
istime was zero and either group g contained a prolog, or g was zero
and there was more than one signal group.  (If g was zero and it was
the only signal group, then isgsetframe would be a no-op.)

This was wrong in multiple ways:

- istime may be zero while we are not reading the zeroth skewed frame:
  if there are skewed signals, getframe calls getskewedframe multiple
  times without incrementing istime.  This meant skewing was broken in
  combination with a prolog, and skewing was also broken in
  combination with multi-group records.

- We may be reading the zeroth skewed frame of a given signal file
  while istime is nonzero: in a multi-segment record, istime is the
  elapsed time from the start of the first segment, not from the start
  of the current segment.  This meant prologs were not handled in the
  second and subsequent segments of a multi-segment record.

What we want instead is to call isgsetframe, before reading from a
given signal group, if that signal group has a prolog and we have
never yet called isgsetframe for that group.