setgvmode: reinstate handling of WFDB_GVPAD.
WFDB_GVPAD was introduced in WFDB 10.4.0 and could be specified as an
argument to setgvmode (along with WFDB_LOWRES or WFDB_HIGHRES.)  Some
applications used this as a way to avoid explicitly handling invalid
samples.

This was unintentionally broken in WFDB 10.5.3, and since then,
setgvmode has simply ignored the WFDB_GVPAD flag.  Restore this
feature to be consistent with the documentation and application
expectations.