isgsetframe: don't break deskewing at start of segment.
isgsetframe is called directly by getskewedframe in some
circumstances.  getskewedframe should not modify dsbi, which is used
by the higher-level getframe.

This fixes the handling of a multi-segment record that contains both
prologs and skews.  (In single-segment records, this was not an issue
because in single-segment records, isgsetframe was only called from
getskewedframe at the very start of the record - and getframe works
around that case specifically.)

However, when isgsetframe is called by isgsettime, we do want to reset
dsbi.  Therefore, do this in isgsettime instead.