signal.5: add documentation of FLAC formats.
Document the format of FLAC signal files produced and consumed by
WFDB.

This is not a complete and formal specification.  WFDB is somewhat
more conservative in its output than what libFLAC requires, and
libFLAC in turn does not, at present, support the entire FLAC format.
However, this document is intended as a guide for what should be
considered a valid WFDB-FLAC signal file.