[WIP] lcheck: test reading and writing FLAC signals.
If FLAC support is available, we want to test that the library can
both read and write signals in FLAC format.  However, the compressed
result is not guaranteed to be identical for different versions of
WFDB, different versions of libFLAC, or different compression settings
given by environment variables.  Moreover, we also want to test that
ordinary binary signal I/O still works as well.

To accommodate this, run the signal I/O tests twice: first, copy
record 100s to 100y, then copy 100y to 100z.  If FLAC support is
available, record 100y will use a FLAC format; otherwise it will use
an ordinary binary format.  (This avoids creating yet another set of
variants for the "expected" lcheck output.)  The resulting 100z.dat
and 100z.hea should be identical to the expected files, regardless of
which format is used for the intermediate record.

Additionally (and to avoid adding to the list of URLs accessed by the
test suite), this gives an opportunity to test setinfo(), and test
that getinfo() and putinfo() correctly handle both header files and
info files.