readheader: remove fixed limit of 254 characters per line.
When reading a header file, previously, wfdb_fgets was used, which
meant that if a line was longer than the arbitrary maximum (256 bytes
including the trailing \n and \0), the line would be split and parsed
as if it was two lines.

Instead, use wfdb_getline, to accomodate lines of any length.

readheader can return in numerous error conditions; to make cleanup
simple, free the line buffer in isigclose (which is also where hheader
is freed.)