wave: fix conflicting definitions of ann_popup_active.
The global variable ann_popup_active is defined in annpan.c with an
initial value of -1.  Previously, it was also defined as a "COMMON"
variable in xvwave.h (meaning it was defined, without an initializer,
when compiling xvwave.c.)

Some C compilers and linkers would permit this, allowing a definition
with an initializer in one translation unit to override a definition
without an initializer in another translation unit.  However, this
behavior is non-standard and seems like a bad idea, and some linkers
don't allow it.