install-wave32: numerous changes to support new systems.
On Fedora, there is no need to explicitly install libgcc.  We do need
libtirpc-devel (any architecture will do), and we need flac-devel in
order to use libFLAC.  We also need xset, if the font packages are not
already installed, to add fonts to the running X server.

This script now supports Debian and Debian-like systems as well as
Fedora and Fedora-like systems.  The package dependencies are mostly
the same with different names.

When we install packages for a non-native architecture, we also should
install the same package for the native architecture.  (Neither dnf
nor apt will do this automatically, and if we don't do this, and an
older version of the native package is already installed, then the
installation will fail because the common files don't match.)

For Debian, XView packages are available for several architectures,
although these are untested apart from i386 and armhf.  Compiling for
armhf on arm64 requires using the gcc-arm-linux-gnueabihf
cross-compiler (there is no gcc-multilib package, and gcc doesn't
support -m32.)

If an X server is running, then after installing the font packages,
add their directories to the font path, so that WAVE will work
immediately without restarting the server.